Move models and handler to one file catalystwan/api/builders/feature_profiles/report.py
