+++ catalystwan/api/builders/feature_profiles/other.py
@@ -4,10 +4,9 @@ import logging
 from typing import TYPE_CHECKING, List
 from uuid import UUID
 
-from catalystwan.api.builders.feature_profiles.handler import handle_build_report
+from catalystwan.api.builders.feature_profiles.report import FeatureProfileBuildReport, handle_build_report
 from catalystwan.api.feature_profile_api import OtherFeatureProfileAPI
 from catalystwan.endpoints.configuration.feature_profile.sdwan.other import OtherFeatureProfile
-from catalystwan.models.builders import FeatureProfileBuildReport
 from catalystwan.models.configuration.feature_profile.common import FeatureProfileCreationPayload
 from catalystwan.models.configuration.feature_profile.sdwan.other import AnyOtherParcel
