+++ catalystwan/api/builders/feature_profiles/report.py
@@ -1,12 +1,34 @@
+import functools
 from typing import Any, List, Optional, Tuple, Union
 from uuid import UUID
 
 from pydantic import BaseModel, ConfigDict, Field
 
-from catalystwan.exceptions import ManagerErrorInfo
-from catalystwan.models.configuration.feature_profile.parcel import ParcelType
+from catalystwan.exceptions import ManagerErrorInfo, ManagerHTTPError
+from catalystwan.models.configuration.feature_profile.parcel import AnyParcel, ParcelType
 
 ParcelName = str
+
+
+def handle_build_report(func):
+    """Wrapper to make creating report simple and keep it DRY."""
+
+    @functools.wraps(func)
+    def wrapper(self, profile_uuid: UUID, parcel: AnyParcel, *args, **kwargs) -> Optional[UUID]:
+        try:
+            uuid = func(self, profile_uuid, parcel, *args, **kwargs)
+            self.build_raport.add_created_parcel(parcel.parcel_name, uuid)
+            return uuid
+        except ManagerHTTPError as e:
+            self.build_raport.add_failed_parcel(
+                parcel_name=parcel.parcel_name,
+                parcel_type=parcel._get_parcel_type(),
+                error_info=e.info,
+                request=e.request,
+            )
+            return None
+
+    return wrapper
 
 
 class FailedRequestDetails(BaseModel):
